Squashed commit of the following:
commit 199e8f85244c8eb1c30163d51755570ad86139bb
Author: Aleksa Sarai <cyphar@cyphar.com>
Date:   Mon Jun 20 19:09:57 2022 +1000

    utils: lockfile: avoid stack overflow for lockfile buffer

    There appears to have been some change on openSUSE (likely some new
    hardening flags for builds, or some glibc hardening) such that incorrect
    buffer handling results in a segfault even if the buffer is never
    overflowed.


commit e9303f684b383c24e4b5226eb54c42439afbf27f
Author: kannibalox <kannibalox@gmail.com>
Date:   Tue Jul 4 02:56:29 2023 -0400

    Remove deprecated std::unary/binary_function (#46)

commit 7a3fc2190f015f8a020c6240ae67a97ed8e5673d
Author: kannibalox <kannibalox@gmail.com>
Date:   Tue Jul 4 02:55:42 2023 -0400

    option_parser: fix unqualified calls to std::move (#59)

    https://reviews.llvm.org/rG70b1f6de539867353940d3dcb8b25786d5082d63

commit f9a875b633eb083ec82f89ebeae8678ca424e8b2
Author: Jesse Chan <jc@linux.com>
Date:   Mon Jul 3 16:13:23 2023 -0700

    download_store: ensure that session files land on the disk

    Currently, save() stores session to the disk by:

    - serialize to bencode and write to files with .new ext
    - validate the new files
    - rename new files, removing .new ext

    Technically, this is done to make sure that session files on the
    disk are always valid/uncorrupted. However, if catastrophic events
    happen during save(), such as system crash / power outage, the actual
    content of session files may not have been synced to the disk.

    To make things much worse, rename() is atomic. That means we discarded
    the previous session files without actually landing the new session files
    on the disk, and we got a bunch of empty session files to deal with.

    This commit adds a fsync step to ensure that session files land on the
    disk.

+++ src/utils/lockfile.cc
@@ -77,7 +77,7 @@ Lockfile::try_lock() {
   if (pos == 0) {
     ssize_t len = std::strlen(buf);
     ::snprintf(buf + len, 255 - len, ":+%i\n", ::getpid());
-    ssize_t __attribute__((unused)) result = ::write(fd, buf, std::strlen(buf));
+    int __attribute__((unused)) result = ::write(fd, buf, std::strlen(buf));
   }
 
   ::close(fd);
